% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATAXIC.R
\name{ATAXIC}
\alias{ATAXIC}
\title{ATAXIC}
\usage{
ATAXIC(input_data_exp)
}
\arguments{
\item{input_data_exp}{A dataframe or matrix containing gene expression profiles in single cancer cells, in which the row name is the gene symbol or Entrez gene ID and the column name is the sample ID of singel cancer cells.}
}
\value{
A dataframe with 2 columns:
\item{cell}{Tumor cell to be calculated.}
\item{ATAXIC_score}{The ATAXIC score of each cell.}
}
\description{
\code{ATAXIC} An algorithm  to quantify transcriptomic perturbation heterogeneity in single cancer cells.
}
\details{
This function calculates the transcriptomic perturbation heterogeneity score in single cancer cells based on the standard deviations of the absolute z-scored gene expression values for tens of thousands of genes.
}
\examples{
ATAXIC(input_data_exp)
}
\author{
Qian Liu \href{mailto:qian.liu@stu.cpu.edu.cn}{qian.liu@stu.cpu.edu.cn}, Xiaosheng Wang \href{mailto:xiaosheng.wang@cpu.edu.cn}{xiaosheng.wang@cpu.edu.cn}, Qiqi Lu \href{mailto:3320092074@cpu.edu.cn}{3320092074@cpu.edu.cn}
}
